move delete handler into its own module
